# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/status": {
            "get": {
                "description": "Retrieve all ping statuses from the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statuses"
                ],
                "summary": "Get all ping statuses",
                "responses": {
                    "200": {
                        "description": "List of ping statuses",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new ping status in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statuses"
                ],
                "summary": "Create a new ping status",
                "parameters": [
                    {
                        "description": "Ping Status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PingStatus"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created ping status",
                        "schema": {
                            "$ref": "#/definitions/model.PingStatus"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to create status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/status/{id}": {
            "get": {
                "description": "Retrieve a specific ping status by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statuses"
                ],
                "summary": "Get a ping status by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ping Status ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ping status",
                        "schema": {
                            "$ref": "#/definitions/model.PingStatus"
                        }
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update a ping status by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statuses"
                ],
                "summary": "Update a ping status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ping Status ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ping Status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PingStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated ping status",
                        "schema": {
                            "$ref": "#/definitions/model.PingStatus"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload or ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to update status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a ping status by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statuses"
                ],
                "summary": "Delete a ping status by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ping Status ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to delete status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Link": {
            "type": "object",
            "properties": {
                "href": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "rel": {
                    "type": "string"
                }
            }
        },
        "model.PingStatus": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Link"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "last_success": {
                    "type": "string"
                },
                "ping_time": {
                    "type": "number"
                }
            }
        }
    }
}